Small refactor + fixed tests
