+++ client/src/containers/sidebar/store.ts
@@ -3,7 +3,6 @@ import { atom } from "jotai";
 
 import { FilterQueryParam } from "@/hooks/use-filters";
 
-// export const customWidgetIdAtom = atom<string | null>(null);
 export const sandboxFiltersAtom = atom<FilterQueryParam[]>([]);
 export const sandboxIndicatorAtom = atom<string | null>(null);
 export const sandboxVisualizationAtom = atom<WidgetVisualizationsType | null>(
