+++ shared/schemas/widget.schemas.ts
@@ -5,12 +5,12 @@ export const CreateCustomWidgetSchema = z.object({
   name: z.string(),
   widgetIndicator: z.string(),
   defaultVisualization: z.string(),
-  filters: z.array(SearchWidgetDataFilterSchema),
+  filters: z.array(SearchWidgetDataFilterSchema).optional(),
 });
 
 export const UpdateCustomWidgetSchema = z.object({
   name: z.string().optional(),
   widgetIndicator: z.string().optional(),
   defaultVisualization: z.string().optional(),
-  filters: z.array(SearchWidgetDataFilterSchema),
+  filters: z.array(SearchWidgetDataFilterSchema).optional(),
 });
